Add first aut implementation

+++ snailsoup-backend/src/services/auth/token_claim.rs
@@ -3,6 +3,6 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize)]
 pub struct TokenClaims {
     pub id: String,
-    pub created_at: usize,
-    pub exp: usize,
+    pub created_at: i64,
+    pub exp: i64,
 }
